@@ -2,14 +2,17 @@
   "name": "goit-react-hw-06-phonebook",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://alena-senkiv.github.io/goit-react-hw-06-phonebook/",
+  "homepage": "https://alena-senkiv.github.io/goit-react-hw-06-phonebook",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "modern-normalize": "^1.1.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-icons": "^4.3.1",
     "react-scripts": "5.0.0",
+    "react-toastify": "^8.2.0",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
@@ -36,5 +39,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "prettier": "^2.6.2",
+    "prop-types": "^15.8.1"
   }
 }
